requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-google-genai
langchain-community
langchain-huggingface  # <-- NOVO
langchain-chroma
chromadb
InstructorEmbedding
sentence-transformers
torch
accelerate
ragas
datasets